+++ src/TGUI/ClickableWidget.cpp
@@ -117,8 +117,8 @@ namespace tgui
         if (m_CallbackFunctions[LeftMousePressed].empty() == false)
         {
             m_Callback.trigger = LeftMousePressed;
-            m_Callback.mouse.x = x - getPosition().x;
-            m_Callback.mouse.y = y - getPosition().y;
+            m_Callback.mouse.x = static_cast<int>(x - getPosition().x);
+            m_Callback.mouse.y = static_cast<int>(y - getPosition().y);
             addCallback();
         }
     }
@@ -131,8 +131,8 @@ namespace tgui
         if (m_CallbackFunctions[LeftMouseReleased].empty() == false)
         {
             m_Callback.trigger = LeftMouseReleased;
-            m_Callback.mouse.x = x - getPosition().x;
-            m_Callback.mouse.y = y - getPosition().y;
+            m_Callback.mouse.x = static_cast<int>(x - getPosition().x);
+            m_Callback.mouse.y = static_cast<int>(y - getPosition().y);
             addCallback();
         }
 
@@ -143,8 +143,8 @@ namespace tgui
             if (m_CallbackFunctions[LeftMouseClicked].empty() == false)
             {
                 m_Callback.trigger = LeftMouseClicked;
-                m_Callback.mouse.x = x - getPosition().x;
-                m_Callback.mouse.y = y - getPosition().y;
+                m_Callback.mouse.x = static_cast<int>(x - getPosition().x);
+                m_Callback.mouse.y = static_cast<int>(y - getPosition().y);
                 addCallback();
             }
